refactor(settings): extract Meeting tab into its own component (#332 phase 1 slice 5)
Fifth mechanical slice of the settings monolith decomposition (after
#387, #389, #390, #391). Moves meeting auto-start mode, speaker-
diarization toggle, the wespeaker model installer (download/cancel/
remove), and the app-classifier overrides panel out of
\`settings/+page.svelte\` into \`MeetingTab.svelte\`.

What moved:
- ~16 \`\$state\` declarations (autostart x3, diarization x3,
  diarizer model status x4 + 3 unlisten handles, diarizer remove
  x3, app overrides x6, app defaults x1).
- ~13 IPC handlers across the auto-start, diarization, diarizer-
  model, and app-overrides surfaces.
- The 5-IPC \`Promise.all\` mount-time load.
- The three \`model:download-progress\`/\`-done\`/\`-failed\` listeners
  filtered by the \`wespeaker-resnet34-lm\` id, plus their
  \`onDestroy\` teardown.
- ~225 lines of markup including the diarizer-model-status branch,
  the click-to-confirm Remove flow, the manual-install disclosure,
  and the \`<MeetingAppOverridesPanel>\` mounting.
- The Meeting-specific CSS (.diarizer-installed-details,
  .diarizer-details, .path-code, .link-like, .diarizer-explainer,
  .diarizer-installed-actions) plus the shared card primitives
  (.settings-group, .toggle-row, .select-row, .group-heading,
  etc.) into MeetingTab's scoped style block.

Behavioural delta worth flagging: if the user starts a diarizer
download from the Meeting tab and then switches to a different
tab, the download itself continues (Rust side is unaffected), but
the live progress percentage stops updating because the listeners
are scoped to the tab's lifecycle. On returning to Meeting,
\`loadDiarizerModelStatus\` re-reads the on-disk state — a
completed download immediately flips the UI to "ready" without
the intermediate progress frames. Same shape as the Whisper
picker on the Model tab. Acceptable per #332's lazy-load goal.

Stats:
- \`settings/+page.svelte\` 1592 → 874 (-718) — biggest single-PR
  shrink in the slice series.
- New \`MeetingTab.svelte\` 791 LOC (CSS is ~270 of that).
- Cumulative across slices 1-5: settings page 2428 → 874 (-1554,
  -64%). About tab is the only remaining markup block.

Refs #332, #392.
